fix: ditch react context for Zustand
Installed Zustand for global state management.
Previously, the upload page will re-render when a row is
added/removed/updated.
The reason behind this is the incorrect usage of React context.
A context was placed at the top of the upload page, and the states of
the rows were placed inside the context. This means when a state
changed, everything under the context will get re-rendered.

Currently, only the file rows section get re-rendered when the rows
updated.