bug: added middleware but now 2FA does not enter admin_dashboard
OTP_verification is supposed to be called in 2FA page and it also deletes previous OTPs if successful. Unit testing using Postman shows OTP_verification works but implementation is not working as the API is never called or never successful? Once a valid OTP is given on the 2FA page, user is routed to login page for some reason.