fix: EmailOtp & OtpVerification is fixed
Both EmailOtp and OTPVerification api is now working, bcrypt.compare
should not be used as == is sufficient.

Attempting to reroute admin to home page if sessionID is invalid when
using the components in the dashbaord page. It console.logs that it is
in home but it remains in Daashboard until the refresh?
At least it prevents invalid SessionID from touching anything

Invalid SessionID will prompt the middleware to delete all cookies and
attempt to route to home.